add localization in login page
